refactor(parish-input): split single-file lib.rs into 6 modules
The crate was a single 1990-LOC file. Group symbols by responsibility:

- commands.rs       — Command, FlagSubcommand enums + branch/flag name validators
- intent_types.rs   — IntentKind, PlayerIntent, InputResult
- parser.rs         — parse_system_command, classify_input
- intent_local.rs   — parse_intent_local (keyword-matching pre-pass)
- intent_llm.rs     — parse_intent (async LLM fallback)
- mention.rs        — MentionExtraction, extract_mention

Public API unchanged: lib.rs now declares the modules and re-exports the
same set of pub items. The test module stays at the bottom of lib.rs to
preserve cross-cutting test coverage; private helpers it tests are imported
explicitly. All 116 lib tests pass.
