+++ crates/parish-input/src/lib.rs
@@ -4,840 +4,27 @@
 //! All other input is natural language sent to the LLM for
 //! intent parsing (move, talk, look, interact, examine).
 
-use parish_config::InferenceCategory;
-use parish_inference::AnyClient;
-use parish_types::GameSpeed;
-use parish_types::ParishError;
-use serde::Deserialize;
+mod commands;
+mod intent_llm;
+mod intent_local;
+mod intent_types;
+mod mention;
+mod parser;
 
-/// Sub-command for the `/flag` feature-flag system.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum FlagSubcommand {
-    /// Enable the named flag.
-    Enable(String),
-    /// Disable the named flag.
-    Disable(String),
-    /// List all known flags.
-    List,
-}
-
-/// A system command entered by the player.
-///
-/// System commands use a `/` prefix and control game meta-operations
-/// like saving, loading, and quitting.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Command {
-    /// Freeze all simulation ticks.
-    Pause,
-    /// Unfreeze simulation.
-    Resume,
-    /// Persist state and exit.
-    Quit,
-    /// Manual snapshot to current branch.
-    Save,
-    /// Create a new named save branch.
-    Fork(String),
-    /// Load a named save branch.
-    Load(String),
-    /// List all save branches.
-    Branches,
-    /// Show history of current branch.
-    Log,
-    /// Show current game status.
-    Status,
-    /// Show help text.
-    Help,
-    /// Toggle the Irish pronunciation sidebar.
-    ToggleSidebar,
-    /// Toggle improv craft mode for NPC dialogue.
-    ToggleImprov,
-    /// Show current LLM provider.
-    ShowProvider,
-    /// Change LLM provider at runtime.
-    SetProvider(String),
-    /// Show current model name.
-    ShowModel,
-    /// Change model name at runtime.
-    SetModel(String),
-    /// Show current API key (masked).
-    ShowKey,
-    /// Set API key at runtime.
-    SetKey(String),
-    /// Debug command with optional subcommand.
-    Debug(Option<String>),
-    /// Show the loading spinner for a duration (seconds).
-    Spinner(u64),
-    /// Show cloud provider info.
-    ShowCloud,
-    /// Change cloud provider at runtime.
-    SetCloudProvider(String),
-    /// Show cloud model name.
-    ShowCloudModel,
-    /// Change cloud model at runtime.
-    SetCloudModel(String),
-    /// Show cloud API key (masked).
-    ShowCloudKey,
-    /// Set cloud API key at runtime.
-    SetCloudKey(String),
-    /// Show current game speed.
-    ShowSpeed,
-    /// Set game speed to a named preset.
-    SetSpeed(GameSpeed),
-    /// Invalid speed preset was requested.
-    InvalidSpeed(String),
-    /// Show provider for a specific inference category.
-    ShowCategoryProvider(InferenceCategory),
-    /// Set provider for a specific inference category.
-    SetCategoryProvider(InferenceCategory, String),
-    /// Show model for a specific inference category.
-    ShowCategoryModel(InferenceCategory),
-    /// Set model for a specific inference category.
-    SetCategoryModel(InferenceCategory, String),
-    /// Show API key for a specific inference category (masked).
-    ShowCategoryKey(InferenceCategory),
-    /// Set API key for a specific inference category.
-    SetCategoryKey(InferenceCategory, String),
-    /// Show about / credits information.
-    About,
-    /// Show or change the map tile source. No arg = list sources; arg = switch to it.
-    Map(Option<String>),
-    /// Open the Parish Designer mod editor.
-    Designer,
-    /// Show NPCs at the current location with details.
-    NpcsHere,
-    /// Show detailed time, weather, and season info.
-    Time,
-    /// Wait in place for a number of game minutes, advancing time.
-    Wait(u32),
-    /// Start a fresh new game, resetting world and NPCs.
-    NewGame,
-    /// Manually tick NPC schedules without advancing time.
-    Tick,
-    /// Show or change the UI theme.
-    Theme(Option<String>),
-    /// Show or set whether unexplored map locations are fully revealed.
-    ///
-    /// `None` reports current usage/status text.
-    /// `Some(true)` reveals all unexplored locations.
-    /// `Some(false)` restores normal fog-of-war frontier visibility.
-    Unexplored(Option<bool>),
-    /// Invalid branch name was provided.
-    InvalidBranchName(String),
-    /// Feature flag management (`/flag enable|disable|list <name>`).
-    Flag(FlagSubcommand),
-    /// List all feature flags (alias for `Flag(List)`).
-    Flags,
-    /// Invalid flag name was provided.
-    InvalidFlagName(String),
-}
-
-/// Maximum allowed length for save branch names.
-const MAX_BRANCH_NAME_LEN: usize = 255;
-
-/// Maximum allowed length for feature flag names.
-const MAX_FLAG_NAME_LEN: usize = 64;
-
-/// Validates a save branch name for length and allowed characters.
-///
-/// Branch names may contain alphanumerics, spaces, underscores, and hyphens.
-fn validate_branch_name(name: &str) -> Result<String, String> {
-    if name.len() > MAX_BRANCH_NAME_LEN {
-        return Err(format!(
-            "Branch name too long (max {} characters).",
-            MAX_BRANCH_NAME_LEN
-        ));
-    }
-    if !name
-        .chars()
-        .all(|c| c.is_alphanumeric() || c == '_' || c == '-' || c == ' ')
-    {
-        return Err(
-            "Branch names may only contain letters, numbers, spaces, underscores, and hyphens."
-                .to_string(),
-        );
-    }
-    Ok(name.to_string())
-}
-
-/// Validates a feature flag name for length and allowed characters.
-///
-/// Flag names may contain alphanumerics, hyphens, and underscores only.
-fn validate_flag_name(name: &str) -> Result<String, String> {
-    if name.is_empty() {
-        return Err("Flag name cannot be empty.".to_string());
-    }
-    if name.len() > MAX_FLAG_NAME_LEN {
-        return Err(format!(
-            "Flag name too long (max {} characters).",
-            MAX_FLAG_NAME_LEN
-        ));
-    }
-    if !name
-        .chars()
-        .all(|c| c.is_alphanumeric() || c == '_' || c == '-')
-    {
-        return Err(
-            "Flag names may only contain letters, digits, hyphens, and underscores.".to_string(),
-        );
-    }
-    Ok(name.to_string())
-}
-
-/// The kind of player action parsed from natural language input.
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum IntentKind {
-    /// Move to a location.
-    Move,
-    /// Talk to an NPC.
-    Talk,
-    /// Look around or at something.
-    Look,
-    /// Interact with an object or NPC.
-    Interact,
-    /// Examine something closely.
-    Examine,
-    /// Intent could not be determined.
-    Unknown,
-}
-
-/// A parsed player intent derived from natural language input.
-///
-/// Created by LLM-based intent parsing of the player's raw text.
-#[derive(Debug, Clone)]
-pub struct PlayerIntent {
-    /// The kind of action the player wants to take.
-    pub intent: IntentKind,
-    /// The target of the action (e.g. an NPC name, location, object).
-    pub target: Option<String>,
-    /// Dialogue text if the player is speaking.
-    pub dialogue: Option<String>,
-    /// The original raw input text.
-    pub raw: String,
-}
-
-/// The result of classifying raw player input.
-///
-/// Input is either a system command (prefixed with `/`) or free-form
-/// game input to be parsed by the LLM.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum InputResult {
-    /// A recognized system command.
-    SystemCommand(Command),
-    /// Free-form game input for LLM parsing.
-    GameInput(String),
-}
-
-/// Attempts to parse a system command from raw input.
-///
-/// Returns `Some(Command)` if the input matches a known `/` command,
-/// `None` otherwise.
-pub fn parse_system_command(input: &str) -> Option<Command> {
-    let trimmed = input.trim();
-    let lower = trimmed.to_lowercase();
-
-    if lower == "/pause" {
-        Some(Command::Pause)
-    } else if lower == "/resume" {
-        Some(Command::Resume)
-    } else if lower == "/quit" {
-        Some(Command::Quit)
-    } else if lower == "/save" {
-        Some(Command::Save)
-    } else if lower == "/fork" || lower.starts_with("/fork ") {
-        let name = trimmed
-            .get("/fork".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if name.is_empty() {
-            Some(Command::Help) // bare /fork → show help
-        } else {
-            match validate_branch_name(&name) {
-                Ok(valid) => Some(Command::Fork(valid)),
-                Err(msg) => Some(Command::InvalidBranchName(msg)),
-            }
-        }
-    } else if lower == "/load" || lower.starts_with("/load ") {
-        let name = trimmed
-            .get("/load".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if name.is_empty() {
-            Some(Command::Load(String::new())) // empty string = show save picker
-        } else {
-            match validate_branch_name(&name) {
-                Ok(valid) => Some(Command::Load(valid)),
-                Err(msg) => Some(Command::InvalidBranchName(msg)),
-            }
-        }
-    } else if lower == "/branches" {
-        Some(Command::Branches)
-    } else if lower == "/log" {
-        Some(Command::Log)
-    } else if lower == "/status" {
-        Some(Command::Status)
-    } else if lower == "/help" {
-        Some(Command::Help)
-    } else if lower == "/irish" {
-        Some(Command::ToggleSidebar)
-    } else if lower == "/improv" {
-        Some(Command::ToggleImprov)
-    } else if lower == "/about" {
-        Some(Command::About)
-    } else if lower == "/map" {
-        Some(Command::Map(None))
-    } else if lower.starts_with("/map ") {
-        let arg = trimmed
-            .get("/map ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if arg.is_empty() {
-            Some(Command::Map(None))
-        } else {
-            Some(Command::Map(Some(arg)))
-        }
-    } else if lower == "/designer" {
-        Some(Command::Designer)
-    } else if lower == "/npcs" {
-        Some(Command::NpcsHere)
-    } else if lower == "/time" {
-        Some(Command::Time)
-    } else if lower == "/where" {
-        Some(Command::Status)
-    } else if lower == "/wait" {
-        Some(Command::Wait(15))
-    } else if lower.starts_with("/wait ") {
-        let mins = trimmed[6..].trim().parse::<u32>().unwrap_or(15);
-        Some(Command::Wait(mins))
-    } else if lower == "/new" {
-        Some(Command::NewGame)
-    } else if lower == "/tick" {
-        Some(Command::Tick)
-    } else if lower == "/theme" {
-        Some(Command::Theme(None))
-    } else if lower.starts_with("/theme ") {
-        let arg = trimmed
-            .get("/theme ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if arg.is_empty() {
-            Some(Command::Theme(None))
-        } else {
-            Some(Command::Theme(Some(arg)))
-        }
-    } else if lower == "/unexplored" {
-        Some(Command::Unexplored(None))
-    } else if lower.starts_with("/unexplored ") {
-        let arg = trimmed
-            .get("/unexplored ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_lowercase();
-        match arg.as_str() {
-            "reveal" | "show" | "on" => Some(Command::Unexplored(Some(true))),
-            "hide" | "off" => Some(Command::Unexplored(Some(false))),
-            _ => Some(Command::Unexplored(None)),
-        }
-    } else if let Some(cmd) = parse_category_command(trimmed, &lower) {
-        Some(cmd)
-    } else if lower == "/provider" {
-        Some(Command::ShowProvider)
-    } else if lower.starts_with("/provider ") {
-        let name = trimmed
-            .get("/provider ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if name.is_empty() {
-            Some(Command::ShowProvider)
-        } else {
-            Some(Command::SetProvider(name))
-        }
-    } else if lower == "/model" {
-        Some(Command::ShowModel)
-    } else if lower.starts_with("/model ") {
-        let name = trimmed
-            .get("/model ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if name.is_empty() {
-            Some(Command::ShowModel)
-        } else {
-            Some(Command::SetModel(name))
-        }
-    } else if lower == "/key" {
-        Some(Command::ShowKey)
-    } else if lower.starts_with("/key ") {
-        let value = trimmed
-            .get("/key ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if value.is_empty() {
-            Some(Command::ShowKey)
-        } else {
-            Some(Command::SetKey(value))
-        }
-    } else if lower == "/spinner" {
-        Some(Command::Spinner(30))
-    } else if lower.starts_with("/spinner ") {
-        let secs = trimmed
-            .get("/spinner ".len()..)
-            .unwrap_or("")
-            .trim()
-            .parse::<u64>()
-            .unwrap_or(30);
-        Some(Command::Spinner(secs))
-    } else if lower == "/debug" {
-        Some(Command::Debug(None))
-    } else if lower.starts_with("/debug ") {
-        let sub = trimmed
-            .get("/debug ".len()..)
-            .unwrap_or("")
-            .trim()
-            .to_string();
-        if sub.is_empty() {
-            Some(Command::Debug(None))
-        } else {
-            Some(Command::Debug(Some(sub)))
-        }
-    } else if lower == "/speed" {
-        Some(Command::ShowSpeed)
-    } else if lower.starts_with("/speed ") {
-        let arg = trimmed.get("/speed ".len()..).unwrap_or("").trim();
-        match GameSpeed::from_name(arg) {
-            Some(speed) => Some(Command::SetSpeed(speed)),
-            None if arg.is_empty() => Some(Command::ShowSpeed),
-            None => Some(Command::InvalidSpeed(arg.to_string())),
-        }
-    } else if lower == "/cloud" {
-        Some(Command::ShowCloud)
-    } else if lower.starts_with("/cloud ") {
-        let rest = trimmed.get("/cloud ".len()..).unwrap_or("").trim();
-        let rest_lower = rest.to_lowercase();
-        if rest_lower.starts_with("provider ") {
-            let name = rest
-                .get("provider ".len()..)
-                .unwrap_or("")
-                .trim()
-                .to_string();
-            if name.is_empty() {
-                Some(Command::ShowCloud)
-            } else {
-                Some(Command::SetCloudProvider(name))
-            }
-        } else if rest_lower == "provider" {
-            Some(Command::ShowCloud)
-        } else if rest_lower.starts_with("model ") {
-            let name = rest.get("model ".len()..).unwrap_or("").trim().to_string();
-            if name.is_empty() {
-                Some(Command::ShowCloudModel)
-            } else {
-                Some(Command::SetCloudModel(name))
-            }
-        } else if rest_lower == "model" {
-            Some(Command::ShowCloudModel)
-        } else if rest_lower.starts_with("key ") {
-            let value = rest.get("key ".len()..).unwrap_or("").trim().to_string();
-            if value.is_empty() {
-                Some(Command::ShowCloudKey)
-            } else {
-                Some(Command::SetCloudKey(value))
-            }
-        } else if rest_lower == "key" {
-            Some(Command::ShowCloudKey)
-        } else {
-            Some(Command::ShowCloud)
-        }
-    } else if lower == "/flags" {
-        Some(Command::Flags)
-    } else if lower == "/flag" || lower == "/flag list" {
-        Some(Command::Flag(FlagSubcommand::List))
-    } else if lower.starts_with("/flag ") {
-        let rest = trimmed.get("/flag ".len()..).unwrap_or("").trim();
-        let rest_lower = rest.to_lowercase();
-        if rest_lower.starts_with("enable ") {
-            let name = rest.get("enable ".len()..).unwrap_or("").trim();
-            match validate_flag_name(name) {
-                Ok(valid) => Some(Command::Flag(FlagSubcommand::Enable(valid))),
-                Err(msg) => Some(Command::InvalidFlagName(msg)),
-            }
-        } else if rest_lower.starts_with("disable ") {
-            let name = rest.get("disable ".len()..).unwrap_or("").trim();
-            match validate_flag_name(name) {
-                Ok(valid) => Some(Command::Flag(FlagSubcommand::Disable(valid))),
-                Err(msg) => Some(Command::InvalidFlagName(msg)),
-            }
-        } else if rest_lower == "enable" || rest_lower == "disable" || rest_lower == "list" {
-            Some(Command::Flag(FlagSubcommand::List))
-        } else {
-            // `/flag <name>` without enable/disable — treat as usage error
-            Some(Command::InvalidFlagName(format!(
-                "Unknown flag sub-command '{}'. Use: /flag enable <name>, /flag disable <name>, /flag list",
-                rest
-            )))
-        }
-    } else {
-        None
-    }
-}
-
-/// Parses dot-notation per-category commands like `/model.dialogue`, `/provider.intent`.
-///
-/// Returns `Some(Command)` if the input matches a `/<base>.<category>` pattern
-/// where base is `model`, `provider`, or `key`, and category is `dialogue`,
-/// `simulation`, or `intent`.
-fn parse_category_command(trimmed: &str, lower: &str) -> Option<Command> {
-    for (prefix, show_fn, set_fn) in &[
-        (
-            "/model.",
-            Command::ShowCategoryModel as fn(InferenceCategory) -> Command,
-            Command::SetCategoryModel as fn(InferenceCategory, String) -> Command,
-        ),
-        (
-            "/provider.",
-            Command::ShowCategoryProvider as fn(InferenceCategory) -> Command,
-            Command::SetCategoryProvider as fn(InferenceCategory, String) -> Command,
-        ),
-        (
-            "/key.",
-            Command::ShowCategoryKey as fn(InferenceCategory) -> Command,
-            Command::SetCategoryKey as fn(InferenceCategory, String) -> Command,
-        ),
-    ] {
-        if let Some(rest) = lower.strip_prefix(prefix) {
-            let (cat_str, arg) = match rest.find(' ') {
-                Some(pos) => (&rest[..pos], trimmed[prefix.len() + pos..].trim()),
-                None => (rest, ""),
-            };
-            let category = InferenceCategory::from_name(cat_str)?;
-            if arg.is_empty() {
-                return Some(show_fn(category));
-            } else {
-                return Some(set_fn(category, arg.to_string()));
-            }
-        }
-    }
-    None
-}
-
-/// Classifies raw input as either a system command or game input.
-///
-/// If the input starts with `/` and matches a known command, returns
-/// `InputResult::SystemCommand`. Otherwise returns `InputResult::GameInput`.
-pub fn classify_input(raw: &str) -> InputResult {
-    let trimmed = raw.trim();
-    if let Some(cmd) = parse_system_command(trimmed) {
-        InputResult::SystemCommand(cmd)
-    } else {
-        InputResult::GameInput(trimmed.to_string())
-    }
-}
-
-/// Raw JSON response from the LLM intent parser.
-#[derive(Deserialize)]
-struct IntentResponse {
-    #[serde(default)]
-    intent: Option<IntentKind>,
-    #[serde(default)]
-    target: Option<String>,
-    #[serde(default)]
-    dialogue: Option<String>,
-}
-
-/// The system prompt used for intent parsing.
-const INTENT_SYSTEM_PROMPT: &str = "\
-You are a text adventure input parser. Given the player's natural language input, \
-determine their intent. Respond with valid JSON containing:\n\
-- \"intent\": one of \"move\", \"talk\", \"look\", \"interact\", \"examine\", \"unknown\"\n\
-- \"target\": what the action is directed at (string or null)\n\
-- \"dialogue\": what the player is saying, if talking (string or null)\n\
-\n\
-IMPORTANT: \"move\" is ONLY for when the player expresses a present desire to \
-navigate somewhere (imperative or future intent). Narrative, past-tense, or \
-reflective statements that merely mention a place name are \"talk\", not \"move\".\n\
-\n\
-Examples:\n\
-Input: \"go to the pub\" → {\"intent\": \"move\", \"target\": \"the pub\", \"dialogue\": null}\n\
-Input: \"talk to Mary\" → {\"intent\": \"talk\", \"target\": \"Mary\", \"dialogue\": null}\n\
-Input: \"tell Padraig I saw his cow\" → {\"intent\": \"talk\", \"target\": \"Padraig\", \"dialogue\": \"I saw his cow\"}\n\
-Input: \"look around\" → {\"intent\": \"look\", \"target\": null, \"dialogue\": null}\n\
-Input: \"pick up the stone\" → {\"intent\": \"interact\", \"target\": \"the stone\", \"dialogue\": null}\n\
-Input: \"I came from the coast\" → {\"intent\": \"talk\", \"target\": null, \"dialogue\": \"I came from the coast\"}\n\
-Input: \"I was at the shore yesterday\" → {\"intent\": \"talk\", \"target\": null, \"dialogue\": \"I was at the shore yesterday\"}\n\
-\n\
-Respond ONLY with valid JSON. No explanation.";
-
-/// Attempts to parse intent locally using keyword matching.
-///
-/// Catches common movement and look phrases without requiring an LLM call.
-/// Returns `None` if the input doesn't match any known pattern.
-pub fn parse_intent_local(raw_input: &str) -> Option<PlayerIntent> {
-    let lower = raw_input.trim().to_lowercase();
-
-    // Movement patterns — multi-word phrases checked first (longest match wins),
-    // then single-verb prefixes. Covers common, colloquial, and unusual verbs.
-    let move_phrases = [
-        "make my way to ",
-        "make my way ",
-        "head over to ",
-        "head over ",
-        "pop over to ",
-        "pop over ",
-        "nip to ",
-        "swing by ",
-        "go to ",
-        "walk to ",
-        "head to ",
-        "move to ",
-        "travel to ",
-        "run to ",
-        "jog to ",
-        "dash to ",
-        "hurry to ",
-        "rush to ",
-        "stroll to ",
-        "saunter to ",
-        "mosey to ",
-        "wander to ",
-        "amble to ",
-        "trek to ",
-        "hike to ",
-        "proceed to ",
-        "sprint to ",
-        "march to ",
-        "traipse to ",
-        "meander to ",
-        "trot to ",
-        "stride to ",
-        "creep to ",
-        "sneak to ",
-        "bolt to ",
-        "scramble to ",
-    ];
-
-    // Single-verb prefixes (without "to") — "saunter pub", "go pub", etc.
-    let move_verbs = [
-        "go ",
-        "walk ",
-        "head ",
-        "visit ",
-        "run ",
-        "jog ",
-        "dash ",
-        "hurry ",
-        "rush ",
-        "stroll ",
-        "saunter ",
-        "mosey ",
-        "wander ",
-        "amble ",
-        "trek ",
-        "hike ",
-        "proceed ",
-        "sprint ",
-        "march ",
-        "traipse ",
-        "meander ",
-        "trot ",
-        "stride ",
-        "creep ",
-        "sneak ",
-        "bolt ",
-        "scramble ",
-    ];
-
-    // Try multi-word phrases first for longest-match semantics
-    for prefix in &move_phrases {
-        if let Some(target) = lower.strip_prefix(prefix) {
-            let target = target.trim();
-            if !target.is_empty() {
-                return Some(PlayerIntent {
-                    intent: IntentKind::Move,
-                    target: Some(target.to_string()),
-                    dialogue: None,
-                    raw: raw_input.to_string(),
-                });
-            }
-        }
-    }
-
-    // Then try bare verb + destination
-    for prefix in &move_verbs {
-        if let Some(target) = lower.strip_prefix(prefix) {
-            let target = target.trim();
-            if !target.is_empty() {
-                return Some(PlayerIntent {
-                    intent: IntentKind::Move,
-                    target: Some(target.to_string()),
-                    dialogue: None,
-                    raw: raw_input.to_string(),
-                });
-            }
-        }
-    }
-
-    // Look patterns
-    let look_phrases = ["look", "look around", "l", "examine room", "where am i"];
-    if look_phrases.contains(&lower.as_str()) {
-        return Some(PlayerIntent {
-            intent: IntentKind::Look,
-            target: None,
-            dialogue: None,
-            raw: raw_input.to_string(),
-        });
-    }
-
-    // First-person narrative guard: sentences that begin with a first-person
-    // pronoun are clearly conversational, never navigation commands.  Catching
-    // them here prevents the LLM from extracting a place name mentioned in the
-    // middle of a statement (e.g. "I came from the coast") as a move target.
-    let first_person_prefixes = ["i ", "i'm ", "i've ", "i'd ", "i'll ", "i was ", "i am "];
-    if first_person_prefixes.iter().any(|p| lower.starts_with(p)) || lower == "i" {
-        return Some(PlayerIntent {
-            intent: IntentKind::Talk,
-            target: None,
-            dialogue: Some(raw_input.trim().to_string()),
-            raw: raw_input.to_string(),
-        });
-    }
-
-    None
-}
-
-/// The result of extracting an `@mention` from player input.
-///
-/// Contains the mentioned name and the remaining input text with the
-/// `@mention` stripped out.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct MentionExtraction {
-    /// The name that was mentioned (without the `@` prefix).
-    pub name: String,
-    /// The remaining input text after stripping the mention.
-    pub remaining: String,
-}
-
-/// Extracts an `@mention` from the beginning of player input.
-///
-/// Recognises `@Name` anywhere in input where `@` appears at the start or
-/// after whitespace. The name runs from the `@` until the next punctuation,
-/// double-space, or end of string — so both single-word names (`@Padraig`)
-/// and multi-word names (`@Padraig Darcy`) are supported.
-///
-/// Returns `None` if no valid `@mention` is found.
-///
-/// # Examples
-///
-/// ```
-/// use parish_input::extract_mention;
-///
-/// let result = extract_mention("@Padraig hello there");
-/// assert_eq!(result.unwrap().name, "Padraig");
-///
-/// let result = extract_mention("hello @Padraig");
-/// assert_eq!(result.unwrap().name, "Padraig"); // also matches after whitespace
-///
-/// let result = extract_mention("no mention here");
-/// assert!(result.is_none());
-/// ```
-pub fn extract_mention(raw: &str) -> Option<MentionExtraction> {
-    let trimmed = raw.trim();
-
-    // Find `@` anywhere in the input (at start, or preceded by a space)
-    let at_pos = trimmed.find('@')?;
-    if at_pos > 0 && !trimmed.as_bytes()[at_pos - 1].is_ascii_whitespace() {
-        return None;
-    }
-
-    let rest = &trimmed[at_pos + 1..];
-    if rest.is_empty() || rest.starts_with(' ') {
-        return None;
-    }
-
-    // Name runs until we hit a delimiter that signals end of the name portion.
-    // Find where the name ends. Name = sequence of words where each word
-    // starts with an uppercase letter or is a short connector (e.g., "O'Brien").
-    // Once we hit a word starting with lowercase (and it's not a name particle),
-    // that's the start of the remaining text.
-    let words: Vec<&str> = rest.splitn(20, ' ').collect();
-    let mut name_end = 0;
-
-    for (i, word) in words.iter().enumerate() {
-        let first_char = word.chars().next().unwrap_or(' ');
-        if i == 0 {
-            // First word is always part of the name
-            name_end = 1;
-            continue;
-        }
-        // If word starts with uppercase, it's likely part of the name
-        if first_char.is_uppercase() {
-            name_end = i + 1;
-        } else {
-            break;
-        }
-    }
-
-    let name = words[..name_end].join(" ");
-    // Remaining = text before the @mention + text after the name
-    let before = trimmed[..at_pos].trim();
-    let after = words[name_end..].join(" ");
-    let remaining = match (before.is_empty(), after.trim().is_empty()) {
-        (true, true) => String::new(),
-        (true, false) => after.trim().to_string(),
-        (false, true) => before.to_string(),
-        (false, false) => format!("{} {}", before, after.trim()),
-    };
-
-    if name.is_empty() {
-        return None;
-    }
-
-    Some(MentionExtraction { name, remaining })
-}
-
-/// Parses natural language input into a structured `PlayerIntent`.
-///
-/// First tries local keyword matching for common commands (movement, look).
-/// Falls back to LLM for ambiguous input. If the LLM call fails,
-/// returns `IntentKind::Unknown`.
-pub async fn parse_intent(
-    client: &AnyClient,
-    raw_input: &str,
-    model: &str,
-) -> Result<PlayerIntent, ParishError> {
-    // Try local parsing first — no LLM needed for obvious commands
-    if let Some(intent) = parse_intent_local(raw_input) {
-        return Ok(intent);
-    }
-
-    let result = client
-        .generate_json::<IntentResponse>(model, raw_input, Some(INTENT_SYSTEM_PROMPT), None, None)
-        .await;
-
-    match result {
-        Ok(resp) => Ok(PlayerIntent {
-            intent: resp.intent.unwrap_or(IntentKind::Unknown),
-            target: resp.target,
-            dialogue: resp.dialogue,
-            raw: raw_input.to_string(),
-        }),
-        Err(_) => Ok(PlayerIntent {
-            intent: IntentKind::Unknown,
-            target: None,
-            dialogue: None,
-            raw: raw_input.to_string(),
-        }),
-    }
-}
+pub use commands::{Command, FlagSubcommand};
+pub use intent_llm::parse_intent;
+pub use intent_local::parse_intent_local;
+pub use intent_types::{InputResult, IntentKind, PlayerIntent};
+pub use mention::{MentionExtraction, extract_mention};
+pub use parser::{classify_input, parse_system_command};
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::commands::{validate_branch_name, validate_flag_name};
+    use crate::intent_llm::IntentResponse;
+    use parish_config::InferenceCategory;
+    use parish_types::GameSpeed;
 
     #[test]
     fn test_parse_quit() {
